refactor: complile project details components in a single folder and transfer it in the _components folder
