+++ src/app/(main)/_components/project-details/fake-data.tsx
